+++ src/main/java/ua/com/foxminded/university/dao/GroupDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 import ua.com.foxminded.university.model.Group;
 
 public interface GroupDao extends GenericDao<Group> {
-    public Group getByIdWithDetail(int id);
-    public List<Group> getByCourseId(int curseId);
+    public Group getByIdWithDetail(Long id);
+    public List<Group> getByCourseId(Long curseId);
     public int updateStudents(Group item);
     public Group getByName(String name);
 }
