+++ src/main/java/ua/com/foxminded/university/service/CourseService.java
@@ -3,6 +3,6 @@ package ua.com.foxminded.university.service;
 import ua.com.foxminded.university.model.Course;
 
 public interface CourseService extends GenericService<Course>{
-    public Course getByIdWithDetail(int id);
+    public Course getByIdWithDetail(Long id);
     public int updateGroups(Course item);
 }
