+++ src/main/java/ua/com/foxminded/university/service/GenericService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface GenericService<T> {
     List<T> getAll();
-    T getById(int id);
+    T getById(Long id);
     T insert(T item);
     void update(T item);
-    void delete(int id);
+    void delete(Long id);
 }
