+++ src/main/java/ua/com/foxminded/university/web/CourseController.java
@@ -49,7 +49,7 @@ public class CourseController {
     }
     
     @GetMapping(value = "/{id}")
-    public String show(@PathVariable("id") Integer id, Model model) {
+    public String show(@PathVariable("id") Long id, Model model) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Showing course. id = {}", id);
         }
@@ -59,7 +59,7 @@ public class CourseController {
     }  
     
     @GetMapping(value = "/edit/{id}")
-    public String editForm(@PathVariable("id") Integer id, Model model) {
+    public String editForm(@PathVariable("id") Long id, Model model) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Editing course. id = {}", id);
         }
@@ -106,7 +106,7 @@ public class CourseController {
     }
     
     @GetMapping(value = "/remove/{id}")
-    public String remove(@PathVariable("id") Integer id) {
+    public String remove(@PathVariable("id") Long id) {
         courseService.delete(id);
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Course has been deleted. id = {}", id);
