+++ src/main/java/ua/com/foxminded/university/web/StudentController.java
@@ -49,7 +49,7 @@ public class StudentController {
     }
 
     @GetMapping(value = "/{id}")
-    public String show(@PathVariable("id") Integer id, Model model) {
+    public String show(@PathVariable("id") Long id, Model model) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Showing student. id = {}", id);
         }
@@ -59,7 +59,7 @@ public class StudentController {
     }
 
     @GetMapping(value = "/edit/{id}")
-    public String editForm(@PathVariable("id") Integer id, Model model) {
+    public String editForm(@PathVariable("id") Long id, Model model) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Editing student. id = {}", id);
         }
@@ -110,7 +110,7 @@ public class StudentController {
     }
 
     @GetMapping(value = "/remove/{id}")
-    public String remove(@PathVariable("id") Integer id) {
+    public String remove(@PathVariable("id") Long id) {
         studentService.delete(id);
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Student has been deleted. id = {}", id);
